+++ src/main/java/com/github/raphcal/localserver/index/IndexServer.java
@@ -2,31 +2,30 @@ package com.github.raphcal.localserver.index;
 
 import com.github.raphcal.localserver.LocalServer;
 import java.io.File;
+import java.io.IOException;
 
 /**
- * Serveur d'index.
+ * Directory index server.
+ *
  * @author Raphaël Calabro (ddaeke-github at yahoo.fr)
  */
 public class IndexServer {
-    
-    public static void main(String[] args) {
-		final File root;
-		if (args.length == 1) {
-			root = new File(args[0]);
-			if (!root.exists()) {
-				System.err.println("Chemin invalide : " + root);
-				return;
-			}
-		} else {
-			root = new File("/home/codio/workspace");
-		}
-		
-		final DirectoryIndexHttpServlet servlet = new DirectoryIndexHttpServlet();
-		servlet.setServerRoot(root);
-		
-		final LocalServer localServer = new LocalServer(8787, servlet);
-		localServer.start();
-        System.out.println("Server started on " + localServer.getEndpoint());
-	}
-    
+
+    public static void main(String[] args) throws IOException {
+        final File root;
+        if (args.length == 1) {
+            root = new File(args[0]);
+            if (!root.exists()) {
+                System.err.println("Invalid root path: " + root);
+                return;
+            }
+        } else {
+            root = new File(System.getenv("HOME"));
+        }
+
+        final LocalServer localServer = new LocalServer(8787, new DirectoryIndexHttpServlet(root));
+        localServer.start();
+        System.out.println("Index server started on " + localServer.getEndpoint() + " listing data from " + root.getCanonicalPath());
+    }
+
 }